Ship the quality-gated INT8 pose model (#3)
Replace the deployed FP32 pose asset with a 74.74% smaller browser-tested INT8 ORT runtime, retaining fixed calibration inputs, quality gates, source evidence, and runtime-only CI fetches.

+++ hooks/useAppCharacterImportActions.ts
@@ -165,7 +165,7 @@ export const useAppCharacterImportActions = ({
       recordStudyEvent("image.processing", {
         outcome: canceled ? "canceled" : "error",
         provider: "wasm",
-        model: "fp32",
+        model: "int8",
         inputPixels: pixelBucket(0),
         workingPixels: pixelBucket(0),
         durationMs: Math.round((performance.now() - startedAt) / 500) * 500,
